abstract out compression method
