+++ exporter/opentelemetry-exporter-otlp-proto-http/src/opentelemetry/exporter/otlp/proto/http/metric_exporter/__init__.py
@@ -18,7 +18,8 @@ from time import sleep
 
 from opentelemetry.exporter.otlp.proto.http import Compression
 from opentelemetry.exporter.otlp.proto.http.exporter import (
-    OTLPExporterMixin, DEFAULT_COMPRESSION, DEFAULT_ENDPOINT, DEFAULT_TIMEOUT
+    OTLPExporterMixin, DEFAULT_COMPRESSION, DEFAULT_ENDPOINT, 
+    DEFAULT_TIMEOUT, _expo, _compression_from_env
 )
 from opentelemetry.sdk.metrics._internal.aggregation import Aggregation
 from opentelemetry.proto.collector.metrics.v1.metrics_service_pb2 import (
@@ -67,25 +68,12 @@ from opentelemetry.sdk.metrics.export import (
 from opentelemetry.sdk.resources import Resource as SDKResource
 from opentelemetry.util.re import parse_headers
 
-import backoff
 import requests
 
 _logger = logging.getLogger(__name__)
 
 
 DEFAULT_METRICS_EXPORT_PATH = "v1/metrics"
-
-# Work around API change between backoff 1.x and 2.x. Since 2.0.0 the backoff
-# wait generator API requires a first .send(None) before reading the backoff
-# values from the generator.
-_is_backoff_v2 = next(backoff.expo()) is None
-
-
-def _expo(*args, **kwargs):
-    gen = backoff.expo(*args, **kwargs)
-    if _is_backoff_v2:
-        gen.send(None)
-    return gen
 
 
 class OTLPMetricExporter(
@@ -126,7 +114,7 @@ class OTLPMetricExporter(
                 environ.get(OTEL_EXPORTER_OTLP_TIMEOUT, DEFAULT_TIMEOUT),
             )
         )
-        self._compression = compression or _compression_from_env()
+        self._compression = compression or _compression_from_env(OTEL_EXPORTER_OTLP_METRICS_COMPRESSION)
         self._session = session or requests.Session()
         self._session.headers.update(self._headers)
         self._session.headers.update(
@@ -426,18 +414,6 @@ def get_resource_data(
         )
 
     return resource_data
-
-
-def _compression_from_env() -> Compression:
-    compression = (
-        environ.get(
-            OTEL_EXPORTER_OTLP_METRICS_COMPRESSION,
-            environ.get(OTEL_EXPORTER_OTLP_COMPRESSION, "none"),
-        )
-        .lower()
-        .strip()
-    )
-    return Compression(compression)
 
 
 def _append_metrics_path(endpoint: str) -> str:
