+++ exporter/opentelemetry-exporter-otlp-proto-http/tests/test_proto_log_exporter.py
@@ -161,7 +161,7 @@ class TestOTLPHTTPLogExporter(unittest.TestCase):
         )
 
     @responses.activate
-    @patch("opentelemetry.exporter.otlp.proto.http._log_exporter.backoff")
+    @patch("opentelemetry.exporter.otlp.proto.http.exporter.backoff")
     @patch("opentelemetry.exporter.otlp.proto.http._log_exporter.sleep")
     def test_handles_backoff_v2_api(self, mock_sleep, mock_backoff):
         # In backoff ~= 2.0.0 the first value yielded from expo is None.
